requirements.txt
fastapi==0.88.0
pytest==7.1.3
uvicorn==0.20.0
sqlalchemy==2.0.7
psycopg2-binary~=2.9.3
python-dotenv
pre-commit
geopandas
faker
numpy
passlib 
pyjwt
jose